+++ CHANGELOG.rst
@@ -61,6 +61,7 @@ Improvements
   data exploration.
 - Added a new plugin to load a series of MCA data from file files located in 
   a single directory.
+- Added a new plugin to crop data by either its indices or data values.
 
 
 Bugfixes